crons be

+++ frontend/src/app/crons/components/EditModal.js
@@ -3,7 +3,7 @@ import Modal from "react-bootstrap/Modal";
 import Form from "react-bootstrap/Form";
 import Button from "react-bootstrap/Button";
 import { useDispatch, useSelector } from "react-redux";
-import { select } from "../../../redux/cronsSlice";
+import {select, setModalOpen} from "../../../redux/cronsSlice";
 import 'ace-builds'
 import 'ace-builds/webpack-resolver'
 
@@ -12,33 +12,54 @@ import "ace-builds/src-noconflict/theme-monokai";
 import "ace-builds/src-noconflict/ext-language_tools";
 import {ColorRing} from "react-loader-spinner";
 import CronsApi from "../../../api/crons";
+import Alert from "react-bootstrap/Alert";
 
 const EditModal = () => {
   const dispatch = useDispatch();
-  const cron = useSelector(state => state.crons.selectedCron)
   const token = useSelector((state) => state.auth.token)
-  const loadingCrons = useSelector(state => state.crons.loadingCrons)
+  const {selectedCron: cron, loadingCrons, errors, modalOpen} = useSelector(state => state.crons)
 
   const [command, setCommand] = useState("");
   const [expression, setExpression] = useState("");
+  const [isActive, setIsActive] = useState(false);
+  const [isManagement, setIsManagement] = useState(false);
+  const [errorOpen, setErrorOpen] = useState(false)
 
   useEffect(() => {
     if (cron) {
       setCommand(cron.command || "")
       setExpression(cron.expression)
+      setIsActive(cron.is_active)
+      setIsManagement(cron.is_management)
     }
   }, [cron]);
+  useEffect(() => {setErrorOpen(!!errors)}, [errors])
 
-  return <Modal centered show={!!cron} onHide={() => dispatch(select())}>
+  const clearModal = () => {
+    setCommand("")
+    setExpression("")
+    setIsActive(false)
+    setIsManagement(false)
+  }
+
+  const closeModal = () => {
+    dispatch(select())
+    dispatch(setModalOpen(false))
+    clearModal()
+  }
+  return <Modal centered show={!!cron || modalOpen} onHide={closeModal}>
     <Modal.Header closeButton>
       <Modal.Title>
         <div className="row">
           <div className="col-lg-12 grid-margin stretch-card">
-            Edit { cron?.command }
+            {cron ? 'Edit' : 'Add new cron'} { cron?.command }
           </div>
         </div>
-        <p className="text-muted mb-0">{cron?.description}</p>
-        <p className="text-muted mb-0">Args: {cron?.arguments?.length && cron.arguments.join(" ")}</p>
+        {
+          cron && <>
+            <p className="text-muted mb-0">{cron?.description}</p>
+          </>
+        }
       </Modal.Title>
     </Modal.Header>
     {
@@ -56,6 +77,7 @@ const EditModal = () => {
           dispatch(select())
         }
       }>
+        {errorOpen && <Alert variant="danger" dismissible onClose={() => setErrorOpen(false)}>{errors}</Alert>}
         <Form.Group className="mb-3">
           <Form.Label>Command</Form.Label>
           <Form.Control
@@ -74,20 +96,59 @@ const EditModal = () => {
             onChange={e => setExpression(e.target.value)}
           />
         </Form.Group>
+        <Form.Group className="mb-3">
+          <Form.Label>Is Active?</Form.Label>
+          <Form.Check
+            checked={isActive}
+            type="switch"
+            id="checkbox"
+            label=""
+            onChange={() => {setIsActive(!isActive)}}
+          />
+        </Form.Group>
+        <Form.Group className="mb-3">
+          <Form.Label>Is Management Command?</Form.Label>
+          <Form.Check
+            checked={isManagement}
+            type="switch"
+            id="checkbox-2"
+            label=""
+            onChange={() => {setIsManagement(!isManagement)}}
+          />
+        </Form.Group>
       </Form>
     </Modal.Body>
     }
     <Modal.Footer>
-      <Button variant="danger" className="float-left" onClick={() => dispatch(CronsApi.delete(token, cron.id))}>
-        Delete
-      </Button>
-      <Button variant="secondary" onClick={() => dispatch(select())}>
-        Close
-      </Button>
-      <Button variant="primary" onClick={() => {
-        dispatch(CronsApi.updateCron(token, cron.id, {command: command, expression: expression}))
-        dispatch(select())
-      }}>
+      {
+        cron && <Button variant="danger" className="float-left" onClick={() => dispatch(CronsApi.delete(token, cron.id))}>
+          Delete
+        </Button>
+      }
+      <Button variant="secondary" onClick={closeModal}>Close</Button>
+      <Button
+          variant="primary"
+          onClick={() => {
+            if (cron)
+                dispatch(CronsApi.updateCron(
+                  token,
+                  cron.id,
+                  {
+                    command: command,
+                    expression: expression,
+                    is_active: isActive,
+                    is_management: isManagement,
+                  }
+              ))
+              else dispatch(CronsApi.create(token, {
+                    command: command,
+                    expression: expression,
+                    is_active: isActive,
+                    is_management: isManagement,
+                  }))
+              clearModal()
+          }}
+      >
         Save Changes
       </Button>
     </Modal.Footer>
